+++ CHANGELOG.md
@@ -46,6 +46,7 @@
 * pool: join futures instead of spawning threads in `RelayPool` methods ([Yuki Kishimoto])
 * pool: reduce overhead by maintaining only one atomic reference count for `RelayConnectionStats` and `RelayFiltering` structs ([Yuki Kishimoto])
 * pool: switch to atomic operations for `RelayStatus` ([Yuki Kishimoto])
+* pool: replace `RwLock` with `OnceCell` for `external_notification_sender` ([Yuki Kishimoto])
 * sdk: deprecate `Client::get_events_of` and `Client::get_events_from` methods ([Yuki Kishimoto])
 * sdk: use `Events` instead of `Vec<Event>` in fetch and query methods ([Yuki Kishimoto])
 * sdk: rename `stream_events_of` to `stream_events` ([Yuki Kishimoto])